+++ core/src/main/kotlin/org/pastalab/fray/core/RunContext.kt
@@ -53,7 +53,6 @@ class RunContext(val config: Configuration) {
   var currentThreadId: Long = -1
   var mainThreadId: Long = -1
   var bugFound: Throwable? = null
-  var mainExiting = false
   var nanoTime = TimeUnit.SECONDS.toNanos(1577768400)
   val hashCodeMapper = ReferencedContextManager<Int>({ config.randomnessProvider.nextInt() })
   var forkJoinPool: ForkJoinPool? = null
@@ -187,14 +186,13 @@ class RunContext(val config: Configuration) {
   fun mainExit() {
     val t = Thread.currentThread()
     val context = registeredThreads[t.id]!!
-    mainExiting = true
     while (registeredThreads.any {
       it.value.state != ThreadState.Completed &&
           it.value.state != ThreadState.Created &&
           it.value != context
     }) {
       try {
-        context.state = ThreadState.Runnable
+        context.state = ThreadState.MainExiting
         scheduleNextOperation(true)
       } catch (e: org.pastalab.fray.runtime.TargetTerminateException) {
         // If deadlock detected let's try to unblock one thread and continue.
@@ -219,7 +217,6 @@ class RunContext(val config: Configuration) {
     val t = Thread.currentThread()
     step = 0
     bugFound = null
-    mainExiting = false
     currentThreadId = t.id
     mainThreadId = t.id
     registeredThreads[t.id] = ThreadContext(t, registeredThreads.size, this)
@@ -337,7 +334,7 @@ class RunContext(val config: Configuration) {
   fun threadCompleted(t: Thread) {
     val context = registeredThreads[t.id]!!
     context.isExiting = true
-    monitorEnter(t)
+    monitorEnter(t, true)
     objectNotifyAll(t)
     context.state = ThreadState.Completed
     //    lockManager.threadUnblockedDueToDeadlock(t)
@@ -539,7 +536,7 @@ class RunContext(val config: Configuration) {
   }
 
   fun lockTryLock(lock: Any, canInterrupt: Boolean, timed: Boolean) {
-    lockImpl(lock, false, false, canInterrupt, timed)
+    lockImpl(lock, false, false, canInterrupt, timed, false)
   }
 
   fun lockImpl(
@@ -547,7 +544,8 @@ class RunContext(val config: Configuration) {
       isMonitorLock: Boolean,
       shouldBlock: Boolean,
       canInterrupt: Boolean,
-      timed: Boolean
+      timed: Boolean,
+      shouldRetry: Boolean,
   ) {
     val t = Thread.currentThread().id
     val context = registeredThreads[t]!!
@@ -587,7 +585,11 @@ class RunContext(val config: Configuration) {
       // not want to rely on ReentrantLock. This allows
       // us to pick which Thread to run next if multiple
       // threads hold the same lock.
-      scheduleNextOperation(true)
+      if (shouldRetry) {
+        scheduleNextOperationAndCheckDeadlock(true)
+      } else {
+        scheduleNextOperation(true)
+      }
       if (canInterrupt) {
         context.checkInterrupt()
       }
@@ -599,12 +601,12 @@ class RunContext(val config: Configuration) {
     }
   }
 
-  fun monitorEnter(lock: Any) {
-    lockImpl(lock, true, true, false, false)
+  fun monitorEnter(lock: Any, shouldRetry: Boolean) {
+    lockImpl(lock, true, true, false, false, shouldRetry)
   }
 
   fun lockLock(lock: Any, canInterrupt: Boolean) {
-    lockImpl(lock, false, true, canInterrupt, false)
+    lockImpl(lock, false, true, canInterrupt, false, false)
   }
 
   fun reentrantReadWriteLockInit(
@@ -1009,9 +1011,6 @@ class RunContext(val config: Configuration) {
             .toList()
             .filter { it.state == ThreadState.Runnable }
             .sortedBy { it.thread.id }
-    if (mainExiting && (currentThreadId == mainThreadId || enabledOperations.size > 1)) {
-      enabledOperations = enabledOperations.filter { it.thread.id != mainThreadId }
-    }
 
     // The first empty check will enable timed operations
     if (enabledOperations.isEmpty()) {
@@ -1021,32 +1020,27 @@ class RunContext(val config: Configuration) {
               .toList()
               .filter { it.state == ThreadState.Runnable }
               .sortedBy { it.thread.id }
-      if (mainExiting && (currentThreadId == mainThreadId || enabledOperations.size > 1)) {
-        enabledOperations = enabledOperations.filter { it.thread.id != mainThreadId }
-      }
     }
 
     // The second empty check throws deadlock exceptions.
     if (enabledOperations.isEmpty()) {
-      if (registeredThreads.all { it.value.state == ThreadState.Completed }) {
-        // We are done here, we should go back to the main thread.
+      // If no thread is blocked. We are done. Return to main thread and exit.
+      if (registeredThreads.values.none { it.state == ThreadState.Blocked }) {
         if (currentThreadId != mainThreadId) {
-          registeredThreads[mainThreadId]!!.unblock()
+          registeredThreads[mainThreadId]?.unblock()
         }
         return
-      } else if (!currentThread.isExiting || Thread.currentThread() is HelperThread) {
-        // Deadlock detected
-        val e = org.pastalab.fray.runtime.DeadlockException()
-        reportError(e)
-        throw e
       }
+      val e = DeadlockException()
+      reportError(e)
+      throw e
     }
 
     step += 1
     if (config.executionInfo.maxScheduledStep in 1 ..< step &&
         !currentThread.isExiting &&
         Thread.currentThread() !is HelperThread &&
-        !(mainExiting && currentThreadId == mainThreadId)) {
+        currentThread.state != ThreadState.MainExiting) {
       currentThread.state = ThreadState.Running
       val e = LivenessException()
       reportError(e)